requirements.txt
pandas
wget
streamlit
Pillow